refactor listenerset status to reuse common functions
