+++ pilot/pkg/config/kube/agentgateway/agentgateway_controller.go
@@ -364,26 +364,14 @@ func (c *Controller) buildFinalListenerSetStatus(
 	routeAttachmentsIndex := krt.NewIndex(routeAttachments, "to", func(o *RouteAttachment) []types.NamespacedName {
 		return []types.NamespacedName{o.To}
 	})
-	return krt.NewCollection(
+	return gatewaycommon.FinalListenerSetStatusCollection(
 		listenerSetStatuses,
-		func(
-			ctx krt.HandlerContext, i krt.ObjectWithStatus[*gatewayv1.ListenerSet, gatewayv1.ListenerSetStatus],
-		) *krt.ObjectWithStatus[*gatewayv1.ListenerSet, gatewayv1.ListenerSetStatus] {
-			routes := routeAttachmentsIndex.Fetch(ctx, config.NamespacedName(i.Obj))
-			routesPerListener := map[string]int32{}
-			for _, r := range routes {
-				routesPerListener[r.ListenerName]++
-			}
-			status := i.Status.DeepCopy()
-			for i, l := range status.Listeners {
-				l.AttachedRoutes = routesPerListener[string(l.Name)]
-				status.Listeners[i] = l
-			}
-			return &krt.ObjectWithStatus[*gatewayv1.ListenerSet, gatewayv1.ListenerSetStatus]{
-				Obj:    i.Obj,
-				Status: *status,
-			}
-		}, opts.WithName("ListenerSetFinalStatus")...)
+		func(ctx krt.HandlerContext, obj *gatewayv1.ListenerSet) []*RouteAttachment {
+			return routeAttachmentsIndex.Fetch(ctx, config.NamespacedName(obj))
+		},
+		func(r *RouteAttachment) string { return r.ListenerName },
+		opts,
+	)
 }
 
 func (c *Controller) buildFinalGatewayStatus(
