+++ pilot/pkg/config/kube/gateway/backend_policies.go
@@ -602,7 +602,7 @@ func BackendTrafficPolicyCollection(
 			}
 		}
 		if len(unsupported) > 0 {
-			msg := fmt.Sprintf("Configuration is valid, but Istio does not support the following fields: %v", humanReadableJoin(unsupported))
+			msg := fmt.Sprintf("Configuration is valid, but Istio does not support the following fields: %v", gatewaycommon.HumanReadableJoin(unsupported))
 			conds[string(gw.PolicyConditionAccepted)].message = msg
 		}
