+++ pilot/pkg/config/kube/gateway/conversion.go
@@ -1866,108 +1866,43 @@ func reportGatewayStatus(
 	gs.Conditions = setConditions(obj.Generation, gs.Conditions, gatewayConditions)
 }
 
-func reportListenerSetStatus(
-	r *gatewaycommon.GatewayContext,
-	parentGwObj *k8s.Gateway,
-	obj *k8s.ListenerSet,
-	gs *k8s.ListenerSetStatus,
-	gatewayServices []string,
-	servers []*istio.Server,
-	gatewayErr *ConfigError,
-	listenersValid bool,
-) {
-	internal, _, _, _, warnings, allUsable := r.ResolveGatewayInstances(parentGwObj.Namespace, gatewayServices, servers)
-
-	// Setup initial conditions to the success state. If we encounter errors, we will update this.
-	// We have two status
-	// Accepted: is the configuration valid. Unlike a Gateway, a ListenerSet is only Accepted if at least
-	// one of its listeners is valid; otherwise we report ListenersNotValid.
-	// Programmed: is the data plane "ready" (note: eventually consistent)
-	gatewayConditions := map[string]*condition{
-		string(k8s.GatewayConditionAccepted): {
-			reason:  string(k8s.GatewayReasonAccepted),
-			message: "Resource accepted",
-		},
-		string(k8s.GatewayConditionProgrammed): {
-			reason:  string(k8s.GatewayReasonProgrammed),
-			message: "Resource programmed",
-		},
+func setProgrammedCondition(gatewayConditions map[string]*condition, internal []string, gatewayServices []string, warnings []string, allUsable bool) {
+	programmed := gatewayConditions[string(k8s.GatewayConditionProgrammed)]
+	mapped := map[string]*gatewaycommon.ListenerStatusCondition{
+		string(k8s.GatewayConditionProgrammed): listenerStatusConditionFromCondition(programmed),
 	}
-	var listenersErr *ConfigError
-	if !listenersValid {
-		listenersErr = &ConfigError{
-			Reason:  string(k8s.ListenerSetReasonListenersNotValid),
-			Message: "None of the ListenerSet's listeners are valid",
-		}
-	}
-
-	if gatewayErr != nil {
-		gatewayErr.Message = "Parent not accepted: " + gatewayErr.Message
-		gatewayConditions[string(k8s.GatewayConditionAccepted)].error = gatewayErr
-	} else if listenersErr != nil {
-		gatewayConditions[string(k8s.GatewayConditionAccepted)].error = listenersErr
-	}
-
-	setProgrammedCondition(gatewayConditions, internal, gatewayServices, warnings, allUsable)
-
-	// A ListenerSet with no valid listeners cannot be programmed; this takes precedence over any
-	// address-related programming status.
-	if listenersErr != nil {
-		gatewayConditions[string(k8s.GatewayConditionProgrammed)].error = listenersErr
-	}
-
-	// Prune stale listener status entries whose listener was removed from the spec,
-	// mirroring reportGatewayStatus. ListenerSetCollection threads the previous status
-	// forward to preserve per-listener data across reconciles, but only refreshes entries
-	// for listeners still in the spec. Without this pruning, a removed listener's status
-	// entry is orphaned forever (len(status.Listeners) > len(spec.Listeners)), which also
-	// wedges observedGeneration once the recomputed status stops differing from the stored one.
-	haveListeners := sets.New[k8s.SectionName]()
-	for _, l := range obj.Spec.Listeners {
-		haveListeners.Insert(l.Name)
-	}
-	listeners := make([]k8s.ListenerEntryStatus, 0, len(gs.Listeners))
-	for _, l := range gs.Listeners {
-		if haveListeners.Contains(l.Name) {
-			haveListeners.Delete(l.Name)
-			listeners = append(listeners, l)
-		}
-	}
-	gs.Listeners = listeners
-
-	gs.Conditions = setConditions(obj.Generation, gs.Conditions, gatewayConditions)
+	gatewaycommon.SetProgrammedCondition(mapped, internal, gatewayServices, warnings, allUsable)
+	applyConditionFromListenerStatus(programmed, mapped[string(k8s.GatewayConditionProgrammed)])
 }
 
-func setProgrammedCondition(gatewayConditions map[string]*condition, internal []string, gatewayServices []string, warnings []string, allUsable bool) {
-	if len(internal) > 0 {
-		msg := fmt.Sprintf("Resource programmed, assigned to service(s) %s", humanReadableJoin(internal))
-		gatewayConditions[string(k8s.GatewayConditionProgrammed)].message = msg
+func listenerStatusConditionFromCondition(c *condition) *gatewaycommon.ListenerStatusCondition {
+	if c == nil {
+		return &gatewaycommon.ListenerStatusCondition{}
 	}
+	out := &gatewaycommon.ListenerStatusCondition{
+		Reason:  c.reason,
+		Message: c.message,
+		Status:  c.status,
+		SetOnce: c.setOnce,
+	}
+	if c.error != nil {
+		out.Error = &gatewaycommon.ListenerStatusConfigError{Reason: c.error.Reason, Message: c.error.Message}
+	}
+	return out
+}
 
-	if len(gatewayServices) == 0 {
-		gatewayConditions[string(k8s.GatewayConditionProgrammed)].error = &ConfigError{
-			Reason:  InvalidAddress,
-			Message: "Failed to assign to any requested addresses",
-		}
-	} else if len(warnings) > 0 {
-		var msg string
-		var reason string
-		if len(internal) != 0 {
-			msg = fmt.Sprintf("Assigned to service(s) %s, but failed to assign to all requested addresses: %s",
-				humanReadableJoin(internal), strings.Join(warnings, "; "))
-		} else {
-			msg = fmt.Sprintf("Failed to assign to any requested addresses: %s", strings.Join(warnings, "; "))
-		}
-		if allUsable {
-			reason = string(k8s.GatewayReasonAddressNotAssigned)
-		} else {
-			reason = string(k8s.GatewayReasonAddressNotUsable)
-		}
-		gatewayConditions[string(k8s.GatewayConditionProgrammed)].error = &ConfigError{
-			// TODO: this only checks Service ready, we should also check Deployment ready?
-			Reason:  reason,
-			Message: msg,
-		}
+func applyConditionFromListenerStatus(c *condition, u *gatewaycommon.ListenerStatusCondition) {
+	if c == nil || u == nil {
+		return
+	}
+	c.reason = u.Reason
+	c.message = u.Message
+	c.status = u.Status
+	c.setOnce = u.SetOnce
+	if u.Error != nil {
+		c.error = &ConfigError{Reason: u.Error.Reason, Message: u.Error.Message}
+	} else {
+		c.error = nil
 	}
 }
 
@@ -1993,50 +1928,6 @@ func reportUnmanagedGatewayStatus(
 	status.Addresses = slices.Map(obj.Spec.Addresses, func(e k8s.GatewaySpecAddress) k8s.GatewayStatusAddress {
 		return k8s.GatewayStatusAddress(e)
 	})
-	status.Listeners = nil
-	status.Conditions = setConditions(obj.Generation, status.Conditions, gatewayConditions)
-}
-
-// reportUnsupportedListenerSet reports a status message for a ListenerSet that is not supported
-func reportUnsupportedListenerSet(class string, status *k8s.ListenerSetStatus, obj *k8s.ListenerSet) {
-	gatewayConditions := map[string]*condition{
-		string(k8s.GatewayConditionAccepted): {
-			reason: string(k8s.GatewayReasonAccepted),
-			error: &ConfigError{
-				Reason:  string(k8s.ListenerSetReasonNotAllowed),
-				Message: fmt.Sprintf("The %q GatewayClass does not support ListenerSet", class),
-			},
-		},
-		string(k8s.GatewayConditionProgrammed): {
-			reason: string(k8s.GatewayReasonProgrammed),
-			error: &ConfigError{
-				Reason:  string(k8s.ListenerSetReasonNotAllowed),
-				Message: fmt.Sprintf("The %q GatewayClass does not support ListenerSet", class),
-			},
-		},
-	}
-	status.Listeners = nil
-	status.Conditions = setConditions(obj.Generation, status.Conditions, gatewayConditions)
-}
-
-// reportNotAllowedListenerSet reports a status message for a ListenerSet that is not allowed to be selected
-func reportNotAllowedListenerSet(status *k8s.ListenerSetStatus, obj *k8s.ListenerSet) {
-	gatewayConditions := map[string]*condition{
-		string(k8s.GatewayConditionAccepted): {
-			reason: string(k8s.GatewayReasonAccepted),
-			error: &ConfigError{
-				Reason:  string(k8s.ListenerSetReasonNotAllowed),
-				Message: "The parent Gateway does not allow this reference; check the 'spec.allowedRoutes'",
-			},
-		},
-		string(k8s.GatewayConditionProgrammed): {
-			reason: string(k8s.GatewayReasonProgrammed),
-			error: &ConfigError{
-				Reason:  string(k8s.ListenerSetReasonNotAllowed),
-				Message: "The parent Gateway does not allow this reference; check the 'spec.allowedRoutes'",
-			},
-		},
-	}
 	status.Listeners = nil
 	status.Conditions = setConditions(obj.Generation, status.Conditions, gatewayConditions)
 }
@@ -2178,6 +2069,16 @@ func buildListener(
 		// We don't generate an istio.Server for it.
 		log.Warnf("protocol is %s but no TLS section is defined, skipping listener creation", l.Protocol)
 		server = nil
+		if _, isListenerSet := obj.(*k8s.ListenerSet); isListenerSet {
+			ok = false
+			listenerConditions[string(k8s.ListenerConditionAccepted)].Error = &gatewaycommon.ListenerStatusConfigError{
+				Reason:  string(k8s.ListenerReasonUnsupportedProtocol),
+				Message: fmt.Sprintf("protocol is %s but no TLS section is defined", l.Protocol),
+			}
+			listenerConditions[string(k8s.GatewayConditionProgrammed)].Error = &gatewaycommon.ListenerStatusConfigError{
+				Reason: string(k8s.GatewayReasonInvalid), Message: "Bad TLS configuration",
+			}
+		}
 	}
 
 	updatedStatus := gatewaycommon.ReportListenerCondition(listenerIndex, l, obj, status, listenerConditions)
@@ -2619,19 +2520,6 @@ func namespacesFromSelector(ctx krt.HandlerContext, localNamespace string, names
 	// Ensure stable order
 	sort.Strings(namespaces)
 	return namespaces
-}
-
-func humanReadableJoin(ss []string) string {
-	switch len(ss) {
-	case 0:
-		return ""
-	case 1:
-		return ss[0]
-	case 2:
-		return ss[0] + " and " + ss[1]
-	default:
-		return strings.Join(ss[:len(ss)-1], ", ") + ", and " + ss[len(ss)-1]
-	}
 }
 
 // NamespaceNameLabel represents that label added automatically to namespaces is newer Kubernetes clusters
