+++ pilot/pkg/config/kube/gateway/testdata/listenerset-hostname-conflict.status.yaml.golden
@@ -65,11 +65,6 @@ status:
       reason: HostnameConflict
       status: "False"
       type: Programmed
-    - lastTransitionTime: fake
-      message: No errors found
-      reason: ResolvedRefs
-      status: "True"
-      type: ResolvedRefs
     name: hostname-conflict-with-gateway-listener
     supportedKinds:
     - group: gateway.networking.k8s.io
@@ -144,11 +139,6 @@ status:
       reason: HostnameConflict
       status: "False"
       type: Programmed
-    - lastTransitionTime: fake
-      message: No errors found
-      reason: ResolvedRefs
-      status: "True"
-      type: ResolvedRefs
     name: hostname-conflict-with-gateway-listener
     supportedKinds:
     - group: gateway.networking.k8s.io
@@ -223,11 +213,6 @@ status:
       reason: HostnameConflict
       status: "False"
       type: Programmed
-    - lastTransitionTime: fake
-      message: No errors found
-      reason: ResolvedRefs
-      status: "True"
-      type: ResolvedRefs
     name: hostname-conflict-with-listener-set-listener
     supportedKinds:
     - group: gateway.networking.k8s.io
@@ -274,11 +259,6 @@ status:
       reason: HostnameConflict
       status: "False"
       type: Programmed
-    - lastTransitionTime: fake
-      message: No errors found
-      reason: ResolvedRefs
-      status: "True"
-      type: ResolvedRefs
     name: hostname-conflict-with-listener-set-listener
     supportedKinds:
     - group: gateway.networking.k8s.io
