+++ pilot/pkg/config/kube/gateway/testdata/listenerset-https-missing-tls.status.yaml.golden
@@ -20,9 +20,9 @@ status:
   - attachedRoutes: 0
     conditions:
     - lastTransitionTime: fake
-      message: No errors found
-      reason: Accepted
-      status: "True"
+      message: protocol is HTTPS but no TLS section is defined
+      reason: UnsupportedProtocol
+      status: "False"
       type: Accepted
     - lastTransitionTime: fake
       message: No errors found
@@ -30,9 +30,9 @@ status:
       status: "False"
       type: Conflicted
     - lastTransitionTime: fake
-      message: No errors found
-      reason: Programmed
-      status: "True"
+      message: Bad TLS configuration
+      reason: Invalid
+      status: "False"
       type: Programmed
     - lastTransitionTime: fake
       message: No errors found
