+++ pilot/pkg/config/kube/gateway/testdata/listenerset-protocol-conflict.status.yaml.golden
@@ -65,11 +65,6 @@ status:
       reason: ProtocolConflict
       status: "False"
       type: Programmed
-    - lastTransitionTime: fake
-      message: No errors found
-      reason: ResolvedRefs
-      status: "True"
-      type: ResolvedRefs
     name: protocol-conflict-with-gateway-listener
     supportedKinds:
     - group: gateway.networking.k8s.io
@@ -142,11 +137,6 @@ status:
       reason: ProtocolConflict
       status: "False"
       type: Programmed
-    - lastTransitionTime: fake
-      message: No errors found
-      reason: ResolvedRefs
-      status: "True"
-      type: ResolvedRefs
     name: protocol-conflict-with-gateway-listener
     supportedKinds:
     - group: gateway.networking.k8s.io
@@ -219,11 +209,6 @@ status:
       reason: ProtocolConflict
       status: "False"
       type: Programmed
-    - lastTransitionTime: fake
-      message: No errors found
-      reason: ResolvedRefs
-      status: "True"
-      type: ResolvedRefs
     name: protocol-conflict-with-listener-set-listener
     supportedKinds:
     - group: gateway.networking.k8s.io
@@ -268,11 +253,6 @@ status:
       reason: ProtocolConflict
       status: "False"
       type: Programmed
-    - lastTransitionTime: fake
-      message: No errors found
-      reason: ResolvedRefs
-      status: "True"
-      type: ResolvedRefs
     name: protocol-conflict-with-listener-set-listener
     supportedKinds:
     - group: gateway.networking.k8s.io
